requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.0
torch>=1.12
PyYAML>=6.0
tqdm>=4.66
rich>=13.0
validators>=0.20
beautifulsoup4>=4.10
lxml>=4.9
requests>=2.28
matplotlib>=3.5
